ignore l2 keystones notifications in tests
when we broadcast notifications for new l2 keystones, there is no guarantee when those broadcasts will be sent out since it is async.  we had a test that would create new keystones, then write an api rpc request, then wait for a response.  occasionally, the response would be the notifications for new l2 keystones that were created earlier in the tests.  we can safely ignore those in this test.