ExternalHeaderMode rebase
Added an "ExternalHeaderMode" for TBC, which disables P2P and allows an external manager to feed in headers to track lightweight consensus.

Code using ExternalHeaderMode can optionally provide a different genesis block and height/diff offsets, to start tracking consensus from a particular block that isn't the true genesis block. If these are set, commands that return height/cumulative difficulty will be adjusted at the TBC level, and the underlying database will not be aware of these offsets.

In ExternalHeaderMode, commands for anything but getting headers (tip, header(s) by hash/height) are disabled.

In ExternalHeaderMode, the code managing the TBC instance has the ability to remove headers from TBC's knowledge, but must provide a known tip to be considered canonical after the removal of headers to ensure deterministic state in the event that header removal results in TBC having to run a tie-breaker which could choose a different tip than upstream expects.

ExternalHeaderMode passes through results from the database for adding and removing headers for the upstream manager to deal with, rather than handling locally with P2P activities as TBC normally does.

Added an optional upstream state id which op-geth can use to track the EVM block which external header mode TBC's knowledge represents.

This upstream state id must live inside TBC rather than op-geth itself in order to ensure that updates to this state id are atomic with changes to the underlying TBC state database, otherwise it would be possible for op-geth to update TBC header state and unexpectedly exit, and then op-geth would be unable to reconcile TBC's state with its own.