add queuing system for bitcoin broadcasts
add a db table btc_transaction_broadcast_request that will store broadcast requests.  when someone wants to broadcast a tx, insert into this table (ignoring duplicates).  have a few workers that read from this table and attempt to broadcast transactions to electrumx.

when a worker "checks out" a broadcast request, it updates the last_broadcast_attempt_at column

when a worker confirms a broadcast, it updates the broadcast_at column

wait 10 minutes before processing same tx again if not confirmed