Fix disconnects caused by ReadDeadline.
Go's ReadDealine has surprising side effects that caused a functional
connection to be terminated prior to actually expiring. By setting the
ReadDeadline to 0 the peer read never times out. This is ok because
there are a bunch of other commands being written that will cause the
read to fail if they do not complete. Simply rely on the chatter to
cause a connection close instead of trying to be clever and detect which
error type the close path received.