change paradigm from checking for missing childrent to keeping track of tip via electrs
prior, to solve for forks we would try to fill in "missing children".  the way we (I) had written it, there is no way to guarantee that we will get to the tip; if there were siblings at that height then we would not check for missing blocks.

a better way to do it it to check for the tip of the btc chain to change; when that changes, update the btc chain until we get to a known block