tbc: fix bug in keystone database and bring back performance (#394)
* Introduce flushing strategy to the database.

We can estimate pretty accurately which transactions need a compaction
and which don't. The issue at hand is that very small transactions may
not flush quickly enough leading to reading of stale data.

This may need some additional tinkering but try this for the time being.

* lint goop

* Fix typo. (#402)

Not sure how this snuck into main.

* typo

* typpo

* FUCK YOU LEVELDB

* Fix prints and remove XXX

* Yikes fix a bad bug in keystone indexer.

We MUST delete the cache in the database function and NOT use continue
to skip over records we don't like. This led to continues growth of the
keystone cache and improper flushing.

Tests aren't passing and need to be investigated.

* typo

* Leave not for maitainers how we reindex the last block always

* Leave note for maitainers how we reindex the last block always

* fixed keystone specific tests

* use maps package for clone

* revert keystone index hash back so tests pass

* Dedup next block from all indexers so that we can call it elsewhere as well.

* More dedup of code and simplification

* Moar smol

* Oops! fix a bug in processKeystones; explanation in code

* made keystone testing more robust in fork tests

* Add some block tests

* added old fork tests back

* linter crap

* make keystone fork tests a bit more organized

---------

Co-authored-by: AL-CT <antoniomartins@protonmail.com>