Disregard keystones that are too far in the future (#510)
* disregard keystones that are too far in the future

if someone broadcasts an invalid keystone, specifically one that with a very high l2 block number, all finalities for lower keystones will be incorrec.

add a configuration for a "baseline l2 block" to calculate the expected now + 10 minute block height.  if we have keystones above that, do not include them in finality calculations.

* fix typo