keep all data needed for btc finality in one table (#519)
* keep data needed for btc finality all in one table

prior, we were calculating finality at request/query-time.  this is a relatively-expensive (and complex) calculation. now, store the "lowest btc block height" in the l2_keystones table.  this allows us to return finality with no joins at request/query-time.  this is all done during bitcoin block processing.

our finality lookup for a keystone is, for lack of a better term, now "constant-time" lookup. as the l2_keystone_abrev_hash is indexed (and that's our lookup key).

* fixed missing return

* make sure we ignore after invalid keystones

* fix lowest published

* fix lint

* less than or equal

* set updated_at

* remove confusing comment

* on delete and update