handle tx and deadlock better when updating l2 keystone finality
when updating l2 keystone finality, it is possible that many bfgs will
try to update the l2 keystone finality at the same time.  this may lead
to an error "deadlock detected" in postgres.  this occurrs when waiting
on a resource for too long, the transaction is aborted.

this, in itself, is ok, and it is suggested retrying the transaction
as the resource that it was waiting for may be available.

this PR does two things:
* updates keystone finalities in a single tx
* retries up to 5 times before erroring