ensure updating keystones doesn't happen for every block all of the time
prior, we would unconditionally update the lowest l2 keystone (the
finality) _BEFORE_ we would error if a block was seen.  on startup, this
is fine and expected, but afterwards we should not be updating keystones
for blocks that we have seen already as this has already happened.

add a check to prevent this