do not let initial walk of the btc chain block following walks
to track the bitcoin chain, we "walk" it whenever the tip changes.  on
startup, we walk to the "starting block", afterwards we only walk blocks
that we haven't seen.

this is fine, however the initial "walk" blocks the subsequent walks. so
we have to walk to the starting block before we start handling new
blocks.  with a higher "starting block", this finishes quickly, but is
still subject to the height of the "starting block".

going forward, perform the initial walk concurrently with watching for
new blocks.

so we would:
* walk the chain on startup to the starting block
* upon tip change, walk back to the first known block (handling new
  blocks and re-orgs)

these can run concurrently with the added mutex