fix fragile test: TestKeystoneFinalityInheritance
fix fragile test by adding a retry to sending finality http requests in
tests.  I believe the root cause of this is within the handling of
responses between gozer <-> tbc, but the correct solution to that will
be needed to be debated internally.

for now, to get us past the "fragile test" hump, just retry the request
that contains that issue if it fails

fixes #605