popm: use single privatekey in popminer (#600)
* Print balance of pop miner at startup\nFixes #211

* Print pop miner pub and address in keygen

* Make the wallet a bit smarter and easier to use.

- Implement proper key derivation from provided path in either uints or
  string.
- Make zuul use standard private keys

There are some other nits that went in as well.

* Simplify keygen to generate user friendly keys.

It now always creates JSON blobs for the various types and it should be
a simple copy and paste into the various tools for usage.

* Use same private key for eth and btc in pop miner.

Code was doing weird stuff with derived keys which made it hard for
users to interact with. Drop all that nonsense and use a single key for
everything.

* Fix mutex madness in Vinz Clortho

* Print pubkey bytes in keygen