localnode: add small section for updating a node (#624)
* added small section for updating a node

added a small section for updating a node.  this section simply tells a user to pull the latest and re-run the `gen-files.sh` script as well as the appropriate compose file.  there _shouldn't_ be anything else that is needed; as localnode will contain the compatible versions and data directories and default configs.

* fixed pr feedback

* add note about following v1

* feedback