tests: in mock op-geth, do not return an error if the websocket was closed (#631)
in mock op-geth, do not return an error if the websocket was closed

if websocket was closed, we will receive an EOF error when trying to read from it.  we don't have a great way to record closed web sockets, or delete them, in tests without a large change, so for now just handle it like this.

fixes #630