goreleaser: temporarily remove docker attestations (#638)
Disable Docker attestations in GoReleaser. With the current setup, this
results in indexes being created for each architecture image, which
cannot be used to then create a Docker cross-platform manifest, which
will cause build failure.

In the future, we can consider switching to using ko.build with
GoReleaser to build Docker images, which properly supports these
features and builds much nicer cross-platform images.