Add a disgusting index abstraction layer.
This abstraction is needed to find the subtle differences between the
various indexers. The surrounding code is 99.9% the same and the copy
paste to add a new indexer is very error prone. Finally bite the bullet
and move on making this code use the same code path.

After this pile of shit has settled we can make a proper indexer
interface.